Fix image path

+++ backend/frame/views.py
@@ -97,18 +97,19 @@ class FrameImageCopyAPI(APIView):
 class FrameImagePrintAPI(APIView):
 
     def post(self, request, *args, **kwargs):
-        photo_data = request.FILES['photo']
-        if photo_data:            
-            filename = request.data.get('filename')
-            file_path = os.path.join(settings.BASE_DIR, '../app/public/photo_saved/', filename)
-            if os.path.exists(file_path) and os.path.isfile(file_path):
-                os.remove(file_path)
-            with open(file_path, 'wb') as f:
-                for chunk in photo_data.chunks():
-                    f.write(chunk)
-            return Response({
-                'photo_url': f'/photo_saved/{filename}' if photo_data else None
-            }, status=status.HTTP_201_CREATED)
+        if request.method == 'POST' and 'photo' in request.FILES:
+            photo_data = request.FILES['photo']
+            if photo_data:            
+                filename = photo_data.name
+                file_path = os.path.join(settings.BASE_DIR, '../app/public/photo_saved/', filename)
+                if os.path.exists(file_path) and os.path.isfile(file_path):
+                    os.remove(file_path)
+                with open(file_path, 'wb') as f:
+                    for chunk in photo_data.chunks():
+                        f.write(chunk)
+                return Response({
+                    'photo_url': f'/photo_saved/{filename}' if photo_data else None
+                }, status=status.HTTP_201_CREATED)
         return Response({
             'error': 'Photo data is required'
         }, status=status.HTTP_400_BAD_REQUEST)
